컨플릭트 컨플릭트 해결하면서 실수한 부분 고침

+++ app/src/main/java/com/ajou/capstone_design_freitag/API/RESTAPI.java
@@ -175,7 +175,7 @@ public class RESTAPI {
         User.getUserinstance().setUserID(user.getUserID());
         user.setPoint(jsonObject.getInt("userPoint"));
         User.getUserinstance().setPoint(user.getPoint());
-        user.setAccuracy(jsonObject.getDouble("userAccuracy"));
+        user.setAccuracy((float) jsonObject.getDouble("userAccuracy"));
         User.getUserinstance().setAccuracy(user.getAccuracy());
         //level임의로
         user.setLevel("starter");
